requirements.txt
Flask==2.1.2
google-cloud-bigquery==2.34.2